+++ p2panda-spaces/src/group.rs
@@ -1,0 +1,3 @@
+// SPDX-License-Identifier: MIT OR Apache-2.0
+
+pub struct Group {}
