+++ p2panda-spaces/src/traits.rs
@@ -1,0 +1,9 @@
+// SPDX-License-Identifier: MIT OR Apache-2.0
+
+pub trait Forge<M> {
+    type Error;
+
+    fn forge(&self, args: ForgeArgs) -> Result<M, Self::Error>;
+}
+
+pub struct ForgeArgs {}
